// src/app/shared/model/LTKM/LtkmAttrContent.Model.ts
export class LtkmAttrContent {
    LtkmCustAttrContentId: number;
    AttrCode: string;
    AttrName: string;
    AttrInputType: string;
    AttrValue: string;
    MasterCode: string;
    Descr: string;
    IsMandatory: boolean;
    DefaultValue: string;
    AttrGroup: string;
    constructor() {
    }
}